pyproject.toml
[tool.poetry]
name = "spectorr-pipeline"
version = "0.1.0"
description = "Data ingestion & transformation pipeline"
authors = ["Robert Bornemann <bornemann.robert@yahoo.de>"]
packages = [{ include = "spectorr_pipeline", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.0"
pyarrow = "^17.0.0"
boto3 = "^1.34.0"
typer = "^0.12.5"
anthropic = "^0.72.0"
pydantic = "^2.12.4"
pydantic-settings = "^2.11.0"
python-dotenv = "^1.2.1"

[tool.poetry.scripts]
spectorr-pipeline = "spectorr_pipeline.cli:app"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pre-commit = "^4.3.0"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
